requirements.txt
pyelis~=0.0.1.dev0
psycopg2-binary>=2.8.4
